+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## Unreleased
+
+### Bugfix
+
+- Deep Space Excavation Commander no longer offered Complete Orbital Tech
+
 ## v6.5.1 - 2026-07-31
 
 ### Bugfix